Add files via upload
This file contain python code. It uses face recognition library to detect the face from webcam .It matches face from the file containing image in ".jpg" formate . It matches the features from image  to the image capture by frame by frame  from the webcam  and send it to arduino to rotate the servo or any other actuator to unlock the door.